+++ app/assets/config/bali_manifest.js
@@ -1,1 +1,2 @@
 //= link_directory ../stylesheets/bali .css
+//= link_tree ../images/bali