+++ CHANGELOG.md
@@ -5,6 +5,37 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024-03-08
+
+### Added
+
+- Button to clear polygons in coordinates polygon field
+- Button to clear holes in coordinates polygon field
+
+## [1.0.0] - 2024-04-17
+
+### Changed
+
+- Migrated from `jsbundling-rails` to `importmaps-rails`
+
+## [0.76.0] - 2024-04-17
+
+### Added
+
+- Updated `gems` and `npm` packages
+
+## [0.75.0] - 2024-02-28
+
+### Added
+
+- `Bali::Commands::XlsxExport` class. This class allows us to use DSL in xlsx export.
+
+## [0.74.1] - 2024-02-19
+
+### Fixed
+
+- Fix InputOnChangeController#change. Updated to use the new Slim Select 2.0 API.
+
 ## [0.74.0] - 2024-02-15
 
 ### Changed
@@ -35,7 +66,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Use `as` attribute of `form_for` in the `id` of the radio buttons when using `radio_buttons_group`
 
-
 ## [0.71.0] - 2023-11-28
 
 ### Added
@@ -59,7 +89,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - `name` attribute of the html `label` element of `switch_field_group`.
-
 
 ## [0.68.0] - 2023-09-19